Move `distiset` to root
